+++ src/wifi/model/wifi-phy.cc
@@ -143,6 +143,15 @@ WifiPhy::GetTypeId()
                           BooleanValue(false),
                           MakeBooleanAccessor(&WifiPhy::SetFixedPhyBand, &WifiPhy::HasFixedPhyBand),
                           MakeBooleanChecker())
+            .AddAttribute(
+                "MaxRadioBw",
+                "The maximum width supported by the radio. It is not possible to configure an "
+                "operating channel with a total width larger than this value. A value of 0 means "
+                "no restriction.",
+                TypeId::ATTR_GET | TypeId::ATTR_CONSTRUCT, // prevent setting after construction
+                MHzValue(MHz_t{0}),
+                MakeMHzAccessor(&WifiPhy::m_maxRadioBw),
+                MakeMHzChecker())
             .AddAttribute("RxSensitivity",
                           "The energy of a received signal should be higher than "
                           "this threshold for the PHY to detect the signal. "
@@ -1291,6 +1300,14 @@ WifiPhy::DoChannelSwitch()
                    });
     m_operatingChannel.Set(segments, m_standard);
     m_operatingChannel.SetPrimary20Index(std::get<3>(m_channelSettings.front()));
+
+    // check that the channel width is supported
+    if (const auto chWidth = GetChannelWidth();
+        (m_maxRadioBw != MHz_t{0}) && (chWidth > m_maxRadioBw))
+    {
+        NS_ABORT_MSG("Attempting to set a " << chWidth << " channel on a station only supporting "
+                                            << m_maxRadioBw << " operation");
+    }
 
     if (changingPhyBand)
     {